Eric Sunshine eliminated signed versus unsigned compilation warning from aws/awsmled.cpp.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@15764 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/aws/awsmled.cpp
@@ -378,7 +378,7 @@ bool awsMultiLineEdit::HandleEvent (iEvent &Event)
       if (composer->HandleKey (eventData, Char, 
 	sizeof (Char) / sizeof (utf32_char), &composedSize) != csComposeNoChar)
       {
-	for (size_t n = 0; n < composedSize; n++)
+	for (size_t n = 0; n < (size_t)composedSize; n++)
 	  InsertChar (Char[n]);
       }
       //InsertChar (Event.Key.Char);
